refactor: split guards utils from types
